Feature/ordercancelation

+++ src/infraestructure/repositories/mongo-products.ts
@@ -59,6 +59,7 @@ export class MongoProductRepository implements IProductRepository {
     });
   }
   async findById(id: string): Promise<Product> {
+    console.log(id)
     const product = await ProductModel.findById(id).exec();
     if (!product) {
       throw new Error("Product not found");
@@ -108,6 +109,11 @@ export class MongoProductRepository implements IProductRepository {
     } as unknown as IProduct);
   }
   async delete(id: string): Promise<void> {
+    const inventory = await InventoryModel.findOne({ productId: id }).exec();
+    console.log(inventory)
+    if (inventory.reservedStock >0){
+      throw new Error ("cannot delete product with reserved stock");
+    }
     const result = await ProductModel.deleteOne({ _id: id }).exec();
     if (result.deletedCount === 0) {
       throw new Error("Product not found or already deleted");
